# docs/openapi/crm-v1-openapi.yaml
openapi: 3.0.3
info:
  title: Canicula CRM V1 API
  version: 0.1.0
  description: Project-based key account CRM V1 interface contract.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /api/payments:
    get:
      tags: [Payments]
      summary: List payments
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: account_id
          in: query
          schema:
            type: integer
            format: int64
        - name: opportunity_id
          in: query
          schema:
            type: integer
            format: int64
        - name: contract_id
          in: query
          schema:
            type: integer
            format: int64
        - name: receivable_plan_id
          in: query
          schema:
            type: integer
            format: int64
        - name: payment_status
          in: query
          schema:
            type: string
        - name: payment_method
          in: query
          schema:
            type: string
        - name: owner_user_id
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment list.
    post:
      tags: [Payments]
      summary: Create payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Payment detail.
  /api/payments/{paymentId}:
    get:
      tags: [Payments]
      summary: Get payment detail
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment detail.
    patch:
      tags: [Payments]
      summary: Update payment
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Payment detail.
  /api/payments/{paymentId}/confirm:
    post:
      tags: [Payments]
      summary: Confirm payment amount
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Payment detail.
  /api/payments/{paymentId}/exception:
    post:
      tags: [Payments]
      summary: Register payment exception
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Payment detail.
  /api/payments/{paymentId}/refund:
    post:
      tags: [Payments]
      summary: Mark payment as refunded
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Payment detail.
...
  /api/attachments:
    get:
      tags: [Attachments]
      summary: List attachments by business object
      parameters:
        - name: object_type
          in: query
          required: true
          schema:
            type: string
        - name: object_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Attachment list.
    post:
      tags: [Attachments]
      summary: Create attachment metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Attachment metadata.
  /api/attachments/upload:
    post:
      tags: [Attachments]
      summary: Upload attachment file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [object_type, object_id, file]
              properties:
                object_type:
                  type: string
                object_id:
                  type: integer
                  format: int64
                file:
                  type: string
                  format: binary
                file_type:
                  type: string
                remark:
                  type: string
      responses:
        '200':
          description: Uploaded attachment metadata.
  /api/attachments/{attachmentId}:
    delete:
      tags: [Attachments]
      summary: Delete attachment metadata
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deletion result.
  /api/attachments/{attachmentId}/download:
    get:
      tags: [Attachments]
      summary: Download uploaded attachment file
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Attachment binary stream.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...